refactor: extract contacts bottomsheet to its own component to clean up its impl
